+++ CHANGELOG.md
@@ -3,9 +3,11 @@
 ### Enhancements:
 * docs/guides/*: Added some basic guides on how to manage different resources with dependencies
 * docs/resources/subaccount.md: Add group_ids to example
-* provider/resources/index.go: Invoke a pause, validate state, and ensure all indexes are deleted on tear down
-* client/models.go: Extend out `Transforms` type to contain `queries` for JQ Transforms
 * docs/resources/view.md: Add an example JQ Transform
+* docs/resources/objectgroup.md: Update live events examples to reflect gcp additions
+* provider/resources/index.go: Invoke a pause, validate state, and ensure all indexes are deleted on tear down
+* provider/resources/objectgroup.go: Enable `live_events_gcp`, rename `live_events` -> `live_events_aws`
+* client/models.go: Extend out `Transforms` type to contain `queries` for JQ Transforms
 
 ### Bug Fixes:
 * provider/resources/objectgroup.go: Fixed updates in place for `options` on apply